Add files via upload
This is the submission of the given assignment to enhance our API development skills. i have done four assignments, where the first was about creating the xml and json files for the student's information. then the xquery files to retrieve the name and ages of the student and other for retrieving the subject. other project was about creating API using PHP to manage a list of products. the third project was an online store.